Move File test to File
